apps/admin/backend: Set test timezone to AKST
Set the timezone for tests to Alaska time so that report printing
snapshots can have a deterministic timezone across dev env and CI. I
picked Alaska because it's close to our production timezones
(importantly, on the same side of UTC) but doesn't
match any of our dev timezones. That way it's easy to tell that the
setting is working and it's not just using our local dev env timezones.

I also changed the tests to mock the time using the local system
timezone rather than a UTC timestamp so that the actual mocked date
doesn't change and is easier to reason about (the date written in the test
matches the date in the snapshot).