lint: wrap `interpret::Error` with a `Box` to make clippy happy
Without the `Box`, the containing `Result` always has to reserve enough space for the `Error` type which could be kinda large (~170 bytes). Wrapping it in a `Box` means that we only need enough space for a pointer, at the cost of needing to follow that pointer when we access the error.