Rust lint cleanup (#4674)
* ci: update docker images

The main change is to update Rust from 1.72.0 to 1.76.0.

* lint: use `const` when we can

* lint: use `#[must_use]` where clippy tells us to

* lint: inline `format!` interpolated identifiers

* lint: allow various lints where required

I'd like to have a way to automatically apply `#[allow(clippy::unwrap_used)]` for all `#[cfg(test)]`, but I'm not sure how to do it.

* lint: set workspace-level lint rules and inherit them in crates

* lint: remove `bacon.toml` since lint config is now global

* lint: wrap `interpret::Error` with a `Box` to make clippy happy

Without the `Box`, the containing `Result` always has to reserve enough space for the `Error` type which could be kinda large (~170 bytes). Wrapping it in a `Box` means that we only need enough space for a pointer, at the cost of needing to follow that pointer when we access the error.

* lint: fix various lints flagged by clippy

- rename some structs et al to avoid using the module name as a prefix or suffix
- remove unneeded `to_owned`/`to_string`/`to_vec`/`clone` calls
- don't do exact comparison of floating point values
- don't check JS object types and then force cast; proceed on cast success
- pass method directly to `map` rather than a closure calling that method
- consolidate identical `match` arm bodies
- don't do `for a in b.iter()`; just use `for a in b`
- opportunistically use `thiserror::Error` for `BallotPageTimingMarkMetadataError`

* lint: ignore `clippy::result_large_err` rather than boxing

We aren't going to have many of these, so I don't think the memory issues are really relevant.

* ci: add `lint` step to crates

Calls `cargo clippy`.

* ci: stop ignoring `ballot-interpreter` as a crate

It's true that much of the code is exercised by the JS wrapper, but that package won't run e.g. `cargo clippy`.

* lint: fix or silence lints in `controllerd`