refactor: replace `canvas` with `@napi-rs/canvas`
Note that this is totally broken. We used `ImageData` from `canvas` in such a way that we used it to represent either grayscale (1 channel) or rgba (4 channels). This may be non-standard and seems to be unsupported by `@napi-rs/canvas`. This will likely require a lot of changes to make it work.