chore(image-utils): explicitly support only RGBA images
While it's possible to create a `ImageData` from `canvas` with a `Uint8ClampedArray` that only has one byte per pixel (a grayscale image), none of the APIs for working with the `ImageData` actually support that, so we always have to convert to RGBA anyway. Now that most of the `image-data` modules are gone, this is simpler to enforce. Going forward, any actual image manipulation should probably be done in Rust, or with a fairly narrow scope in JS and/or using the `canvas` APIs.