feat(image-utils): add `fromGrayScale` function
We may have a difficult time getting completely away from converting from a grayscale pixel buffer to an `ImageData` in JS, so this adds a function to do that. I specifically made it so that it takes separate pixels/width/height arguments to avoid trying to represent such a thing as an `ImageData`.