chore(image-deps): remove unused parts (#4683)
* chore(image-utils): remove unused module `dither`

* chore(image-utils): remove unused module `rotate`

* chore(image-utils): remove unused module `binarize`

* chore(image-utils): remove unused module `diff`

* chore(image-utils): remove unused module `otsu`

* chore(image-utils): remove unused module `geometry`

* chore(image-utils): remove unused module `debug`

* chore(image-utils): remove unused module `outline`

* chore(image-utils): remove unused module `count`

* chore(image-utils): remove function `toGrayscale`

Also removes the one use of the function in the test for `saveSheetImage`; a use that was not very valuable.

* chore(image-utils): explicitly support only RGBA images

While it's possible to create a `ImageData` from `canvas` with a `Uint8ClampedArray` that only has one byte per pixel (a grayscale image), none of the APIs for working with the `ImageData` actually support that, so we always have to convert to RGBA anyway. Now that most of the `image-data` modules are gone, this is simpler to enforce. Going forward, any actual image manipulation should probably be done in Rust, or with a fairly narrow scope in JS and/or using the `canvas` APIs.

* fix: remove remaining `toRgba` uses

Inlines an RGBA conversion in the one place we need it.

* fix(converter-nh-accuvote): remove outdated property from mock

* fix(custom): ensure input is grayscale and fix RGBA conversion

* feat(image-utils): add `fromGrayScale` function

We may have a difficult time getting completely away from converting from a grayscale pixel buffer to an `ImageData` in JS, so this adds a function to do that. I specifically made it so that it takes separate pixels/width/height arguments to avoid trying to represent such a thing as an `ImageData`.