Ballot rendering and VX default template
Adds the ballot rendering engine built on top of the Renderer
abstraction. Also adds the VX default template, demonstrating how a
template interacts with this engine. Some components which will be
common to all templates (and are integral to the logic of rendering a
ballot/extracting grid layouts) are pulled out into their own module.