Add jest-image-matcher
This allows us to compare saved ballot fixtures with generated ballot
fixtures and prevent them from getting out of date. We previously tried
comparing the PDF content as text but it was pretty brittle and broke
with these changes, so I switched to a visual comparison, which is more
meaningful anyway.