fix(auth): parse all TLV responses properly (#4700)
* fix(auth): parse all TLV responses properly

Rather than trimming metadata by ignoring fixed-length prefixes and suffixes, we now properly parse the TLV structures. Additionally, this changes the `parseTlv` function to expect the data to be entirely consumed by the TLV structure. If it isn't, use `parseTlvPartial` instead, which returns a remainder.

* fix(auth): revert back to expecting 0x00 uncompressed

I'm not sure why I thought I was seeing 0x01, but I verified that the cert data is not gzipped and that the cert info is 0x00 uncompressed.

* refactor(auth): swap the order of TLV and remainder

https://github.com/votingworks/vxsuite/pull/4700/files#r1535883117