HMPB ballot rendering using Playwright/Chromium (#4703)
* Set up Renderer abstraction to wrap headless Chromium browser

The Playwright-specific bits are pulled out into a separate file in
order to later add support for a regular browser renderer for
interactive render previews.

* Ballot rendering and VX default template

Adds the ballot rendering engine built on top of the Renderer
abstraction. Also adds the VX default template, demonstrating how a
template interacts with this engine. Some components which will be
common to all templates (and are integral to the logic of rendering a
ballot/extracting grid layouts) are pulled out into their own module.

* Add gap parameter to column layout algorithm

This is used in the new ballot template.

* Add SVG assets for ballot template

Moved these into JSX since it was the easiest way to render them. Will
probably need a different solution eventually for custom assets, but
this works for now.

* Add browser preview functionality

(This preview script could use some cleaning up, and the files could be organized a bit better, but for now just leaving it a bit messy. Will revisit in a cleanup PR afterwards.)

* Ballot marking

Modeled after hmpb/layout/mark_ballot

* Add helper to concat pdfs

We used to concatenate laid out ballot documents before creating PDFs
(e.g. for test decks). It's too complicated to try to concatenate
RenderDocuments, since they are living DOM with stylesheets and such.
It's much easier to do concatenation with the resulting static PDFs.

* Add jest-image-matcher

This allows us to compare saved ballot fixtures with generated ballot
fixtures and prevent them from getting out of date. We previously tried
comparing the PDF content as text but it was pretty brittle and broke
with these changes, so I switched to a visual comparison, which is more
meaningful anyway.

* Update HMPB fixture generation logic

- Use next ballot rendering
- Separate the generation of the ballots/election def (which now requires significant resources and is only used for saving the ballot PDFs) from generating the fixture source data (which is used in tests that rely on the fixtures)
- Enable loading of the saved fixtures from file by exposing the file
  paths (previously, tests would just regen the fixture data, since it was
cheap)

* Regenerate HMPB fixtures

* VxDesign: Integrate next ballot renderer

* VxDesign: Integrate next ballot rendering for test deck creation

* Update test deck tests

I had to change some of the fixtures used for a few reasons:
- To use elections that already have gridLayouts instead of generating
  them on demand
- To prevent the PDF snapshots from getting too huge

* Save new test deck PDF snapshot images

* Update ballot-interpreter tests

* Update WIA crop interpretation snapshots

* Update dependencies

* Remove old render_ballot code and assets

* Bump test timeout for HMPB fixtures

* Update ballot rendering mocks in apps/design/backend tests

* Add plumbing to put translated election strings in election definition

* Regenerate HMPB fixtures

* Update test snapshots

Due to change in election hash from translated election strings

* PR clarifications

* Speed up HMPB fixture tests in CI

Pare down the tests to the minimum checks that should catch most
inconsistencies.

* Bump test timeout