pdi-scanner: Detect disconnect and send event (#4773)
* pdi-scanner: Detect disconnect and send event

Detect when the scanner is disconnected (e.g. unplugged) while polling
the scanner for incoming events. This manifests as
`rusb_async::Error:Stall`. Propagate this error back up to the main
thread and send a disconnect error event to stdout.

We also break from the loop that is detecting events when we receive an
error, which causes the scanner_to_host channel to close, so we handle
that case in the main thread by dropping the client (allowing the user
to reconnect if desired).

* Update libs/pdi-scanner/src/scanner.rs

Co-authored-by: Brian Donovan <1938+eventualbuddha@users.noreply.github.com>

---------

Co-authored-by: Brian Donovan <1938+eventualbuddha@users.noreply.github.com>