pdi-scanner: Differentiate command/response/event JSON objects
Since errors can be sent in response to commands or as events, it makes
life easier if we can differentiate them. While I was at it, I also
changed the command JSON format to clarify that it is a command. This makes logs of scanner command/response cycles much clearer.