pdi-scanner: Support command queueing in TS client (#4771)
The Rust code already queues commands and processes them one a time, so
we can adapt the TS client to match this behavior and receive responses
in the order that requests were sent. That way, the consuming code in
the state machine doesn't need to worry about waiting for previous
requests to complete before sending a new request (which can be an issue
when transitioning from a state that polls scanner status to another
state that sends an accept command or the like).