apps/scan: Detect cover open/closed in PDI state machine
Add a root level listener that will listen for scanner events all the
time to detect cover open events. This also simplifies states that were
only adding a listener to detect disconnect events, since that can be
handled by the root listener. In order to hit this listener, we need to
make sure that no other event listener handle has a default transition,
so that we always fall through to the root listener (this is why we remove the "unexpected
event" handlers elsewhere and move it only to the root listener).

In addition, we need to use the scanner status to catch cover open on
connection, since there's no event sent.

Overall, while this approach seems a little complicated compared to our
previous scanner integrations, since we can't poll scanner status while
waiting for ballots (due to aforementioned race conditions), events seem
like the way to go.