apps/scan: Stay in disconnected state when attempting reconnection
Previously, I took a shortcut by reusing the `connecting` state after a
disconnection to attempt reconnection. This caused a flicker in the
frontend since we show an empty screen during the `connecting` state,
while we show an error message in the `disconnected` state. Instead,
this commit adds reconnection logic within the `disconnected` state to
ensure we continuously show the error message until we have reconnected.