apps/scan: Poll for PDI scanner cover closed status (#4797)
I noticed that it was possible for the coverClosed event to be emitted before the state machine had transitioned to the coverOpen state (e.g. if you open the cover slightly, remove a jammed ballot, and then immediately close it), in which case we get stuck in the coverOpen state. Polling the scanner status seems like a more reliable way to exit this state.

Note that I'm not concerned about missing the coverOpen event since that listener is registered at the root of the state machine.

We could conceivably also register the coverClosed event handler at the root, but I am worried that it's a bit heavy handed to say that we should transition to waitingForBallot at any point if we receive a coverClosed event (rather than only when we're already in the coverOpen state).