libs/ballot-interpreter: Bugfix for binarized scanner images
If the image is binarized by the scanner before reaching the
interpreter, the Otsu threshold will be computed as 0 (pure black). We
were only accepting bottom timing marks whose luminance value was less
than the threshold, which was impossible with a threshold of 0. To fix
this, we allow matching the threshold exactly.