apps/scan: Reject second ballot inserted during accept
Previously, we were waiting for the scanner to no longer have any document in the scanner to declare accepting complete. It's possible to trick the scanner by feeding a second ballot in while the motors are running to accept the first ballot, which results in a document in the front of the scanner, but not the rear.

We handle this case with a few updates:
- Only declare accepting complete when the rear is cleared
- If the front is covered before the rear is cleared (very unlikely, but possible), go back to the warning to remove the second ballot before deciding whether the first ballot was succesfully accepted