apps/scan: Fix pdictl process leak
When the scanner is disconnected, we create a new client each time we
reconnect, which spawns a new pdictl process. This process is not killed
when the previous client goes out of scope.

To fix the immediate problem, we exit the client explicitly before
creating a new one. There may be a cleaner design for the client that
will solve this in a better way, but I wanted to fix the immediate
problem as a starting point.