Bugfix: In error state, check error code before exiting
Previously, this state would invoke the `exit` command immediately (while
also checking the error code in the `always` clause). This wasn't a
problem before, since calling `exit` on the client was ok, but with the
new code from the previous commit that calls `exit` in the
`disconnected` state, this resulted in `exit` being called twice, which
caused a fatal error.