apps/scan: Fix pdictl process leak (#4820)
* apps/scan: Fix pdictl process leak

When the scanner is disconnected, we create a new client each time we
reconnect, which spawns a new pdictl process. This process is not killed
when the previous client goes out of scope.

To fix the immediate problem, we exit the client explicitly before
creating a new one. There may be a cleaner design for the client that
will solve this in a better way, but I wanted to fix the immediate
problem as a starting point.

* Bugfix: In error state, check error code before exiting

Previously, this state would invoke the `exit` command immediately (while
also checking the error code in the `always` clause). This wasn't a
problem before, since calling `exit` on the client was ok, but with the
new code from the previous commit that calls `exit` in the
`disconnected` state, this resulted in `exit` being called twice, which
caused a fatal error.