libs/pdi-scanner: Increase timeout so long ballots can eject
Previously, when trying to eject a long (22in) ballot from the back of the
scanner to the front of the scanner, the subsequent scanner status
command would time out. We bump this timeout from 1s to 2s since that
seems to be reliably give the scanner enough time to finish ejecting a
long ballot.