apps/scan: Reuse scanner client to reconnect when scanner is disconnected
Instead of exiting pdictl and creating a new client (thus spawning a new
pdictl), we can now safely reuse the same client when we get the
`'disconnected'` error code, because we know it means that the scanner
is actually disconnected (rather than potentially some other reason that
we can't contact the scanner - e.g. it crashed).