libs/pdi-scanner: Reset `scan_in_progress` flag on errors/disconnections
Previously, we only reset it when we got an `EndScan` event. This wasn't
a noticeable issue because we were previously closing `pdictl` whenever
the scanner was disconnected. But now we need to reset the
`scan_in_progress` flag more precisely - whenever a scan stops, whether
that be the result of success, error, or disconnection.