apps/scan: Add root event listener to new client after error recovery
The state machine creates an initial client on startup and registers a
root event listener to handle events that occur at any point during the
state machine's lifetime. When recovering from an error, we were not
re-registering that root level event listener on the newly created
scanner client.

This commit also fixes a bug where the entry action to create the new
client was not running before the invoke to call `connect` on that
client. It turns out that you aren't allowed to rely on relative
ordering between entry and invoke.