Improve PDI scanner error handling (#4832)
* libs/pdi-scanner: Add 'exited' error code to indicate pdictl exited

This will differentiate two cases:
- `'disconnected'` - pdictl is running but the scanner is disconnected
- `'exited'` - pdictl has exited

That way, consumers (i.e. the state machine) will know when to
disconnect vs when to create a new client.

* apps/scan: Reuse scanner client to reconnect when scanner is disconnected

Instead of exiting pdictl and creating a new client (thus spawning a new
pdictl), we can now safely reuse the same client when we get the
`'disconnected'` error code, because we know it means that the scanner
is actually disconnected (rather than potentially some other reason that
we can't contact the scanner - e.g. it crashed).

* apps/scan: Handle errors when reconnecting to the PDI scanner

Sometimes, the scanner is in a borked state when we try to reconnect, so
we should handle that error differently, not just keep trying to
reconnect.

* libs/pdi-scanner: Reset `scan_in_progress` flag on errors/disconnections

Previously, we only reset it when we got an `EndScan` event. This wasn't
a noticeable issue because we were previously closing `pdictl` whenever
the scanner was disconnected. But now we need to reset the
`scan_in_progress` flag more precisely - whenever a scan stops, whether
that be the result of success, error, or disconnection.

* apps/scan: Add root event listener to new client after error recovery

The state machine creates an initial client on startup and registers a
root event listener to handle events that occur at any point during the
state machine's lifetime. When recovering from an error, we were not
re-registering that root level event listener on the newly created
scanner client.

This commit also fixes a bug where the entry action to create the new
client was not running before the invoke to call `connect` on that
client. It turns out that you aren't allowed to rely on relative
ordering between entry and invoke.

* logging: Enable debug logs with mockLogger

This helps with debugging state machine tests