apps/scan: Disable scanning while scanner cover is open
If left to its own devices, the PDI scanner will continue trying to scan
ballots while the cover is open. Of course, this doesn't actually work
because the rollers can't grab the paper. This might actually be an
issue if attempting to clean the scanner while the app is running. So we
explicitly disable scanning while the cover is open.