apps/scan: Enable PDI scanner eject pause feature (#4843)
* libs/pdi-scanner: Enable eject pause feature when ejecting to rear

Enables the eject pause feature only for rear ejections (ballot
accepts). Adds the corresponding events for pause/resume.

Also addresses a subtle bug that I've only seen as a result of the eject pause feature. When the scanner pauses, it queues any commands received during the pause. After eject, we poll for scanner status to ensure the eject completed succesfully. Since there is a small gap between when the scanner pauses and when we receive the pause event, our status command would be queued by the scanner and our driver code would timeout the command. Later, when the scanner resumes, it would respond to the command, which we would store as an unhandled packet. On our next status request, we would return that previous status packet, and then store the new status response as an unhandled packet, and so on. Thus, all future status requests would be off by one - receiving the previous scanner status.

To address this, I opted to clear out all unhandled packets that are
command responses (i.e. not events) before sending a new command that
expects a response. I'm open to suggestions if there are other ways to
restructure our driver to make it resilient to this case - this just
seemed like a minimal change that would address the issue.

* apps/scan: Update state machine to use ejectPaused/ejectResumed events

* Clippy