apps/scan: Only listen for scanner events at root
We don't actually want to have multiple active listeners at a time for
scanner events. The listener at the root will send a state machine event
that can be caught in any state. Registering an additional listener
within a state will just create a duplicate state machine event. This
hasn't been causing problems because we've been transitioning away from
the states with listeners before that happens, but it still presents a
risk and is unnecessary.