apps/scan: Set PDI scanner paper length to prevent back-to-back ballots (#4875)
Previously, if you timed it just right, it was possible to trick the PDI
scanner into thinking it was scanning one long ballot by inserting two
ballots in quick succession back to back. In this case, the first ballot
would be ejected out of the rear of the scanner into the ballot box
before the issue would be detected by the interpreter.

To prevent this, we set the max paper length setting provided by the PDI
scanner. After scanning (approximately) this length, the PDI scanner
declares a jam and stops the motors. I found that the optimal length for
stopping the motors soon enough to prevent the first ballot from being
ejected was 0.5in shorter than the actual length of the ballot.