fix(ui): cache modal `appElement` to fix tests
In tests an element with id "root" typically does not exist, so `Modal` falls back to `document.body.firstElementChild`. However, `ReactModal` (under our `Modal`) adds an element with attribute `data-react-modal-body-trap` at the root. So if the `Modal` is ever re-rendered, it'll get a different `appElement` the next time around. This is important because when the modal first appears `ReactModal` adds `aria-hidden` to the `appElement`, and then when the modal goes away it removes the `aria-hidden` attribute. However, if the `appElement` changes between renders then you will end up with the `aria-hidden` attribute sticking around, which means all the `screen.findByRole` calls will fail because they ignore anything inside an element with `aria-hidden`.