apps/design: Remove bubble position and density controls
The new VX ballot template doesn't support these parameters. My guess is
that changing the bubble position will be state-specific and therefore
be better implemented in a different template (rather than trying to
have complex templates that support both bubble positions). Density may
be a feature we want to implement, but if we do decide to keep it, I'd
like to make it a template-specific parameter.