libs/hmpb cleanup (#4882)
* hmpb/render-backend: Move files out of `next` dir

Also create a `preview` dir for browser preview specific files

* Move code from hmpb/layout to hmpb/render-backend

* apps/design: Remove NH custom content feature

We don't support it in the VX ballot template, since we'll instead want
to build a separate NH ballot template. I'd like to come up with a more
general way to add custom parameters for ballot templates, so there's
not much value in keeping this feature around.

* Update libs/ballot-interpreter

* apps/design: Remove bubble position and density controls

The new VX ballot template doesn't support these parameters. My guess is
that changing the bubble position will be state-specific and therefore
be better implemented in a different template (rather than trying to
have complex templates that support both bubble positions). Density may
be a feature we want to implement, but if we do decide to keep it, I'd
like to make it a template-specific parameter.

* Remove libs/hmpb/layout

* Move libs/hmpb/render-backend -> libs/hmpb

Most of these changes are rote, though I also updated the README and
accidentally snuck those changes into this commit.

* Update coverage config