Create new i18next instance for every backend render
Instead of relying on a single global i18next instance and updating all
of its resources every render, create a new instance every render. This
is just a little cleanup to make this flow less confusing. Now each
render more clearly loads the latest translations from the store and
creates a language context to be used by the React components.