Bugfix: Wait for content to render before measuring
When measuring elements, we need to make sure the browser has time to
finish rendering them before taking measurements. I noticed a few cases
where SVGs were slowing down the rendering enough that the measurements
were off. Adding the tiniest sleep, which seems to resolve it (maybe by delaying the event loop slightly?). I don't know of a
good way to wait for all of the elements to render, which would be a
more ideal solution, so going with this fix for now.