libs/hmpb: Wait for HTML to load during rendering
When inserting HTML content into the DOM, we want to wait for it to
finish rendering before doing anything with it (e.g. taking
measurements).

I found the `MutationObserver` API, which registers an observer to call
a callback when various types of changes happen to a DOM node.
Specifically, the one that is useful to us is a change in the node's
children.