apps/design: Export all translations in electionStrings.json
Previously, we just used this file for translations that couldn't be
stored in the CDF election.json. Now, we use it as the main file for
translated election strings, since the plumbing for this file is already
set up. We may need to reconsider this if we want to hash translated
election strings that end up on HMPBs, but this should be workable for
now.