VxDesign: VXF/CDF election export format option (#4977)
* ui: Extract CheckboxButton component from CheckboxGroup

* apps/design: Add CDF checkbox to export screen

Default to serializing elections in VXF unless the checkbox is checked
to use CDF. Note that this applies to all of the exports, since they all
need to serialize the election definition in order to hash it.

* apps/design: Export all translations in electionStrings.json

Previously, we just used this file for translations that couldn't be
stored in the CDF election.json. Now, we use it as the main file for
translated election strings, since the plumbing for this file is already
set up. We may need to reconsider this if we want to hash translated
election strings that end up on HMPBs, but this should be workable for
now.

* Update code to read election strings

No changes to logic here, just clarifying the new situation by renaming things.

* Update tests and fixtures

* Regenerate HMPB fixtures

Election hash change

* Update snapshots that depend on HMPB fixtures