fix(ballot-interpreter): allow interpreting both BMD and HMP ballots
Refs #4659

This will increase the time taken to scan BMD ballots for elections that have a grid layout defined. We can do better by unifying the BMD and HMP interpretation code paths, which I will do later. This is a smaller change that removes the current limitation while I work on the bigger unification.