fix(ballot-interpreter): allow interpreting both BMD and HMPB ballots (#4978)
* test(ballot-interpreter): generate BMD ballots on demand

* refactor(ballot-interpreter): extract ballot helpers

* test(ballot-interpreter): unify configuration and replace more image fixtures

* test(ballot-interpreter): unify configuration and replace more image fixtures

* fix(ballot-interpreter): allow interpreting both BMD and HMP ballots

Refs #4659

This will increase the time taken to scan BMD ballots for elections that have a grid layout defined. We can do better by unifying the BMD and HMP interpretation code paths, which I will do later. This is a smaller change that removes the current limitation while I work on the bigger unification.

* chore: remove support for raw binary encoded QR codes

We haven't used them in a long time, and the only reason this has remained as long as it has is because we previously had a lot of test fixtures that used this encoding.

* test(coverage): bring coverage back up above threshold