refactor(ballot-interpreter): clean up some of the scoring code
At some point after Rust 1.76 the dead code linting became more aggressive, which led to the fields on `ScoredBubbleMark` that we weren't serializing being marked as unused. I removed the fields and, in the process, cleaned up some of the code that finds the best match. Rather than having three variables which are all either `None`/some sentinel value OR `Some`/a real value, we now have one variable that is either `None` when no match has been found or `Some` if one has.