libs/hmpb: Rework fix to wait for HTML to load during rendering (#4982)
I had tried out using a MutationObserver to wait for DOM content to load
after being inserted. However, the same issue occurred again, where some
content was still changing after the observer fired. I think it only
appeared to work by delaying things slightly without actually solving the problem.

Here is a new approach where we use requestAnimationFrame to wait for
the next browser repaint. This seems to work, though similarly to last
time I can't be 100% confident that it's solving the problem, so we'll
need to wait and see.