Add `ballotStrings` field to `Election` type
This will hold the translated election strings from the rest of the election definition (currently, the contents of electionStrings.json). Eventually, we'll probably also want to add the other strings that go on the HMPB (e.g. instructions) so they get hashed as well, so that's why I called it `ballotStrings`.