feat(ballot-interpreter): reject ballots whose timing mark borders are way off
When the timing mark border lines are determined to be more than 25% inset from the image edge we now reject the ballot. This should prevent a variety of situations where we might accidentally infer the wrong set of shapes as the best set of timing marks for a given edge.

To make that inference less likely, this change also is smarter about picking the subset of rectangles we determine is the set of timing marks for a given side. To do that we now favor subsets of rectangles that are closest to the edge we're looking for while also having a favorable count (i.e. closest to expected value).