feat(ballot-interpreter): reject ballots whose timing mark borders are way off (#4999)
* feat(ballot-interpreter): reject ballots whose timing mark borders are way off

When the timing mark border lines are determined to be more than 25% inset from the image edge we now reject the ballot. This should prevent a variety of situations where we might accidentally infer the wrong set of shapes as the best set of timing marks for a given edge.

To make that inference less likely, this change also is smarter about picking the subset of rectangles we determine is the set of timing marks for a given side. To do that we now favor subsets of rectangles that are closest to the edge we're looking for while also having a favorable count (i.e. closest to expected value).

* style: fix issues flagged by `cargo clippy`

* feat(ballot-interpreter): update allowed timing mark inset to 15%

25% was just too much leeway. It's okay to reject ballots a bit more.

* test(ballot-interpreter): restore tests that were accidentally deleted

* fix(ballot-interpreter): use 15% of width for all timing mark inset values

15% for each dimension for letter size is 1.25" L/R and 1.65" T/B, which is probably okay. But for legal size it's 1.25" L/R and 2.1" T/B and for 22" long paper the T/B is 3.3". That is too much. Instead, I think it makes more sense to use the 15% of width value for all sides.

* fix(ballot-interpreter): use 10% inset since it works well with real hardware

I tested it with @arsalansufi's latest images from real hardware and even the worst rotation had the timing marks comfortably within the 10% inset. I updated the images for the regression test that failed with this inset value to center the ballot more.

* fix(converter-nh-accuvote): find timing marks in templates properly

Template images often have a 1" margin around edge, making an allowed timing mark inset of 10% too conservative. For that code path, we now allow using a custom value, which is set to 25% for templates.