fix(ballot-interpreter): handle smudges on timing marks better (#5002)
When a smudge that's connected to a timing mark causes the contours algorithm to report a rectangle containing both the timing mark and the adjacent smudge, we might filter out the timing mark because it's too big but fail to filter out some smaller rectangle that's in the smudge. This smaller rectangle might then be used as the timing mark, leading to incorrect interpretation. This commit ignores timing marks that are too different in size from the others in their edge, letting them be inferred when generating the complete set of timing marks.