fix(ballot-interpreter): fix unmarked write-in scoring with skew
Fixes https://github.com/votingworks/vxsuite/issues/4571

When ballots are skewed the write-in areas are no longer aligned with the image pixel grid. The previous version of scoring the unmarked write-in areas simply calculated the bounds of the area and counted all the black pixels in that area. That had the downside of counting some of the write-in line and/or label or other pre-printed material on the ballot, leading to false positives. This change counts the number of black pixels within the shape defined by the four corner points of the write-in area, i.e. a quadrilateral. This greatly reduces the score when no write-in is present on skewed ballots, leading to much lower rates of false positives.