fix(ballot-interpreter): reject ballots with high skew or rotation  (#5015)
* test(ballot-interpreter): consolidate some boilerplate

* fix(ballot-interpreter): reject ballots with high skew or rotation

While scanning ballots may be rotated (whole ballot image is ~uniformly rotated, i.e. a straightforward transform rather than a distortion) and/or skewed (parts of the ballot experienced stretching due to different parts of the ballot passing through the scanner at different rates). We can deal with some amount of both, but the higher the rotation or skew the more likely we are to read the ballot incorrectly. This caps the allowed ballot rotation at 2° and the allowed skew (as measured by how far away from 90° each corner angle is) at 1°.

* fix(ballot-interpreter): reduce skew false positives

Only error for high rotation or skew if we inferred at least one non-metadata timing mark.

* test(ballot-interpreter): remove ballot that no longer scans

And update the snapshots.