Add `Election.id` field
This ID will be generated by VxDesign and added to the election def (and
thus the ballot hash). That way, we can safely encode this ID on smart
cards (instead of the ballot hash), while ensuring that the smart cards
are configured for the same election as the ballots and the machines.