VxMark/VxMarkScan: Remove ReplaceElectionScreen (#5046)
* apps/mark-scan|mark: Remove ReplaceElectionScreen

This screen is a relic of an era where machines were configured by the
election manager card alone. You could replace the configuration by
inserting a different election manager card. Since that's no longer
possible due to using USBs to configure, this screen is no longer
useful. It's basically a different version of the "invalid card" screen
(which is standard across apps). We rely on that screen instead now.

In order to trigger the invalid card screen (which is already rendered for other invalid card cases), we remove the auth flag
that allows authenticating with an election manager card that
corresponds to a different election.

* Remove auth flag

Remove the `allowElectionManagersToAccessMachinesConfiguredForOtherElections` flag from `InsertedSmartCardAuth`, since it was only used for this use case.

* Update pnpm-lock