fix(basics): improve types of conditional `iter` methods
When it only makes sense to call certain `iter` methods based on the contained type `T`, the better approach is to use an explicit type for `this` in the method definition rather than a ternary type resulting in `unknown` when the type doesn't match. This way actually gives us TypeScript errors and is clearer.