fix(pdi-scanner): use a larger buffer when receiving (#5097)
The PDI scanner doesn't seem to stream as it receives image data from its sensors, so it sends the data all in one go if it can. This reduces the number of incoming transfers for the image data to just one, which should reduce the odds of any image corruption that might happen if e.g. a packet is dropped. It might also make it a tiny bit faster, but I doubt the effect would be noticeable.