perf(scan): optimize writing of scanned images to disk
Previously we were writing both front & back images to disk twice: once before interpretation (as a kind of staging area, I guess?) and once after. This removes the first write (and the subsequent read+decode). In addition, this ensures that the PNG encoding (which is the slowest part of saving the images) happens in parallel.