Strict ballot definition CDF (#5129)
* Remove `vxBallotLabel` extension field from ballot definition CDF

This was a minor improvement that pretty much was only used for showing
the party name for a candidate on VxMark. It's fine to just use the full
name. I considered making it optional, but just removing it makes life
even simpler.

* Use a default for `optionBoundsFromTargetMark` when converting CDF election

* Remove seal custom field from ballot definition CDF

CDF elections will just have no seal shown on screen. Modify the Seal
component to support this case.