perf(scan): optimize writing of scanned images to disk (#5132)
* perf(scan): optimize writing of scanned images to disk

Previously we were writing both front & back images to disk twice: once before interpretation (as a kind of staging area, I guess?) and once after. This removes the first write (and the subsequent read+decode). In addition, this ensures that the PNG encoding (which is the slowest part of saving the images) happens in parallel.

* fix(scan): make central-scan compatible with interpretation changes

Unlike VxScan, VxCentralScan gets its images from the disk (via `scanimage`) so there's no savings to be had here without totally replacing `scanimage`. This just makes the minimal changes needed to load the image data from the paths we were already provided.

* docs(image-utils): remove outdated JSDoc

* test(scan): fix various tests that broke with the last change

* test(ballot-interpreter): get back within expected coverage range

* fix(ballot-interpreter): save the normalized images, not the originals

* fix(scan): clean the data to log with Custom too

* test(scan): work around race conditions due to timing changes

* test(ballot-interpreter): adjust coverage thresholds for deleted code