VxScan: Accept ballots on backend after interpretation
There's no need to wait for the frontend to send an accept command. Note
that I chose to _not_ check auth status and whatnot before accepting.
While we've had that guard in the past, I figured that the time between
starting the scan and accepting is so short that it seems reasonable to
just rely on the check we already do before starting the scan.